Extract sessions into separate files
